raftserver: rename region to route.
