+++ src/raftserver/store/mod.rs
@@ -9,7 +9,7 @@ mod cmd_resp;
 mod store;
 mod peer;
 mod peer_storage;
-mod region;
+mod route;
 mod util;
 
 pub use self::msg::{Msg, SendCh, Callback, call_command};
